pyproject.toml
[tool.poetry]
name = "testlangchain"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.0"
langchain-aws = "^0.2.0"
python-dotenv = "^1.0.1"
langserve = {extras = ["all"], version = "^0.3.0"}
langchain-chroma = "^0.1.4"
langchain-community = "^0.3.0"
langgraph = "^0.2.21"
aidial-sdk = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
